src/main/java/com/example/movement/controller/UserController.java
package com.example.movement.controller;

import com.example.movement.dto.UserDTO;
import com.example.movement.service.UserService;
import jakarta.validation.Valid;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;


@RestController
@RequestMapping("/users")
public class UserController {

    private final UserService userService;

    public UserController(UserService userService) {
        this.userService = userService;
    }

    @PostMapping("/register")
    public ResponseEntity<UserDTO> registerUser(@Valid @RequestBody UserDTO userDTO) {
            return ResponseEntity.ok(userService.registerUser(userDTO));
    }
}
